requirements.txt
fairscale==0.4.6
pandas==1.3.5
pytorch_lightning==1.6.3
PyYAML==6.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cu111
transformers==4.19.2
protobuf==3.20.1